+++ src/corpus/createCorpus.composable.ts
@@ -32,6 +32,7 @@ export default function useCreateCorpus() {
     name: string,
     description: string,
     format: CorpusSourceFormat,
+    language: string,
     textAnnotation?: string,
   ) {
     const configOptions = {
@@ -39,6 +40,7 @@ export default function useCreateCorpus() {
       name: createByLang(name),
       description: createByLang(description),
       format,
+      language,
       textAnnotation,
     };